controllers,
services,
configuration for tables,
components for empployee

+++ Beneficiaries.Solution/Beneficiaries.API/Controllers/BeneficiaryController.cs
@@ -50,7 +50,7 @@ namespace Beneficiaries.API.Controllers
         [HttpDelete("Delete/{id}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public async Task<ActionResult<string>> Delete(double id)
+        public async Task<ActionResult<string>> Delete(Int64 id)
         {
             if (id == 0)
             {
@@ -73,7 +73,7 @@ namespace Beneficiaries.API.Controllers
 
         [HttpGet("{id}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
-        public object GetById(double id)
+        public object GetById(Int64 id)
         {
             if (id == 0)
             {
@@ -87,7 +87,7 @@ namespace Beneficiaries.API.Controllers
 
         [HttpGet("ObtAllXEmployeeId")]
         [ProducesResponseType(StatusCodes.Status200OK)]
-        public async Task<ActionResult<IEnumerable<object>>> ObtAllXEmployeeId(double employeeId)
+        public async Task<ActionResult<IEnumerable<object>>> ObtAllXEmployeeId(Int64 employeeId)
         {
             _logger.LogInformation("Get list x Employee Id");
             var LItems = await _beneficiaryService.ObtAllXEmployeeId(employeeId);
